# ShelterHub/client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Longhl104.ShelterHub",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Pets/shelter/{shelterId}": {
      "get": {
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "shelterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetPetsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPetsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPetsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "shelterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pets/shelter/{shelterId}/paginated": {
      "get": {
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "shelterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "nextToken",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PetStatus"
            }
          },
          {
            "name": "species",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "breed",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaginatedPetsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaginatedPetsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaginatedPetsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pets/{id}": {
      "get": {
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pets/{id}/status": {
      "put": {
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetStatus"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PetStatus"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PetStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pets/{petId}/upload-url": {
      "post": {
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "petId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fileName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contentType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileSizeBytes",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pets/images/download-urls": {
      "post": {
        "tags": [
          "Pets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPetImageDownloadUrlsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPetImageDownloadUrlsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetPetImageDownloadUrlsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PetImageDownloadUrlsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetImageDownloadUrlsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetImageDownloadUrlsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreatePetRequest": {
        "required": [
          "breed",
          "dateOfBirth",
          "description",
          "gender",
          "name",
          "species"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "species": {
            "type": "string",
            "nullable": true
          },
          "breed": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetPaginatedPetsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "pets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pet"
            },
            "nullable": true
          },
          "nextToken": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetPetImageDownloadUrlsRequest": {
        "type": "object",
        "properties": {
          "petRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PetImageDownloadUrlRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetPetsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "pets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pet"
            },
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Pet": {
        "type": "object",
        "properties": {
          "petId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "species": {
            "type": "string",
            "nullable": true
          },
          "breed": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/PetStatus"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "shelterId": {
            "type": "string",
            "format": "uuid"
          },
          "mainImageFileExtension": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PetImageDownloadUrlRequest": {
        "type": "object",
        "properties": {
          "petId": {
            "type": "string",
            "format": "uuid"
          },
          "mainImageFileExtension": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PetImageDownloadUrlsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "petImageUrls": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PetResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "pet": {
            "$ref": "#/components/schemas/Pet"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PetStatus": {
        "enum": [
          "Available",
          "Pending",
          "Adopted",
          "MedicalHold"
        ],
        "type": "string"
      },
      "PresignedUrlResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "presignedUrl": {
            "type": "string",
            "nullable": true
          },
          "key": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}